Migrate users of the changed UA function too

+++ app/src/main/java/tipz/viola/webview/BrowserActivity.kt
@@ -138,7 +138,7 @@ class BrowserActivity : VWebViewActivity() {
             currentPrebuiltUAState = !currentPrebuiltUAState
             webview.setPrebuiltUAMode(
                 view,
-                (if (currentPrebuiltUAState) 1 else 0).toDouble(),
+                if (currentPrebuiltUAState) 1 else 0,
                 false
             )
         } else if (item == R.drawable.new_tab) {
